+++ src/heplers/formatPhoneForBack.js
@@ -1,0 +1,3 @@
+export const formatPhoneForBack = (number) => {
+  return number.replace(/\s|-/g, "");
+};
